👌 Use extend path when calling Qt tools (#52)
This is mainly an issue when using docker images (e.g. in the CI) where DLLs like msvcp140_1.dll are not on the path.
This then causes uic and rcc to crash since they cloud not find the DLLs.

For windows additional DLLs are shipped with shiboken6.

Missing *.so files (E.g. libQt6Core.so.6) might also cause problems on linux, so I preemptively added the lib folders.